+++ src/4.features/text-editor/ui/add-link/add-link.module.scss
@@ -1,0 +1,38 @@
+.wrapper {
+  display: flex;
+  padding: 5px 0;
+  border-radius: 5px;
+  align-items: center;
+  justify-content: flex-start;
+  gap: 20px;
+}
+
+.label {
+  padding: 5px 10px;
+  border-radius: 5px;
+  background: var(--message);
+  display: flex;
+  align-items: center;
+  min-width: 300px;
+  gap: 5px;
+}
+
+.input {
+  background: transparent;
+  flex: 1 1 auto;
+}
+
+.button {
+  padding: 5px;
+  border-radius: 5px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  &:hover {
+    background: var(--background);
+  }
+}
+
+.lucide {
+  color: var(--color);
+}
