--- src/4.features/text-editor/ui/link-portal/link-portal.module.scss
@@ -1,36 +1,0 @@
-.wrapper {
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  background: var(--overlay);
-  z-index: 100;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  background: var(--background);
-  padding: 20px;
-  border-radius: 5px;
-  align-items: center;
-  gap: 20px;
-  min-width: 300px;
-}
-
-.input {
-  padding: 10px;
-  border-radius: 5px;
-}
-
-.button {
-  padding: 5px;
-  border-radius: 5px;
-  background: var(--dark-travis);
-}
