+++ src/4.features/text-editor/ui/text-editor.module.scss
@@ -6,18 +6,22 @@
 
 .buttons {
   position: absolute;
-  top: -25px;
-  background: var(--background);
+  top: -40px;
+  border-radius: 5px;
+  background: var(--message);
+  color: var(--background);
+
+  padding: 5px;
   display: flex;
   align-items: center;
-  gap: 10px;
+  gap: 5px;
 }
 
 .editor {
   padding: 10px 0;
   width: 100%;
   font-size: 20px;
-  max-height: 100px;
+  max-height: 300px;
   overflow: auto;
   word-break: break-word;
 }
@@ -29,22 +33,22 @@
   width: 100%;
   border-radius: 20px;
   background-color: var(--message);
+  color: var(--color);
   padding: 10px 40px;
   font-size: 20px;
 }
 
 .lucide {
-  color: #ffffff;
+  color: var(--color);
 }
 
 .button {
-  background: var(--overlay);
-  padding: 2px;
+  padding: 5px;
   border-radius: 5px;
   display: flex;
   align-items: center;
   justify-content: center;
   &:hover {
-    background: var(--dark-travis);
+    background: var(--background);
   }
 }
